+++ .github/workflows/deploy.yml
@@ -1,6 +1,6 @@
 on:
   push:
-    branches: [step1]
+    branches: [step2]
 
 jobs:
   deploy:
@@ -23,4 +23,4 @@ jobs:
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
           publish_dir: ./build
-          publish_branch: step1-pg
+          publish_branch: step2-pg
