+++ src/components/ProductCard/ProductCard.tsx
@@ -1,8 +1,8 @@
 import { memo } from "react";
 import { ReactComponent as ShoppingCartImg } from "../../assets/icon/shopping-cart.svg";
-import Counter from "../Counter/Counter";
-import ProductImg from "./ProductImg/ProductImg";
-import ProductInfo from "./ProductInfo/ProductInfo";
+import Counter from "../common/Counter/Counter";
+import ProductImg from "../ProductImg/ProductImg";
+import ProductInfo from "../ProductInfo/ProductInfo";
 import useCart from "../../hooks/useCart";
 import Styled from "./ProductCardStyled";
