+++ src/components/ProductInfo/ProductInfoStyled.ts
@@ -2,6 +2,7 @@ import styled from "styled-components";
 
 const Styled = {
   ProductInfo: styled.div`
+    width: 80%;
     display: flex;
     flex-direction: column;
     gap: 4px;
