+++ src/components/common/Counter/Counter.stories.tsx
