+++ src/components/common/Counter/CounterStyled.ts
